Вынес вспомогательные функции в отдельный файл
