+++ helpers.php
@@ -142,32 +142,3 @@ function include_template($name, array $data = []) {
 
     return $result;
 }
-
-function get_dt_range($date): array{
-    date_default_timezone_set('Europe/Moscow');
-    $expiry_date = DateTime::createFromFormat('Y-m-d', $date);
-    $expiry_date->setTime(23, 59, 59);
-    $currentDate = new DateTime();
-    $dt_range = $currentDate->diff($expiry_date);
-
-    $hours = 0;
-    $minutes = 0;
-
-    if (! $dt_range->invert) {
-        $hours = $dt_range->days * 24 + $dt_range->h;
-        $minutes = $dt_range->i;
-    }
-
-    $hours = str_pad($hours, 2, "0", STR_PAD_LEFT);
-    $minutes = str_pad($minutes, 2, "0", STR_PAD_LEFT);
-
-    return [$hours, $minutes];
-}
-
-function xss_protection($string): string{
-    return htmlspecialchars($string);
-}
-
-function price_format($price): string{
-    return number_format(ceil($price),0, '.',' ').' ₽';
-}
