+++ index.php
@@ -1,6 +1,7 @@
 <?php
 require_once('helpers.php');
 require_once('db_connection.php');
+require_once('service_functions.php');
 
 $is_auth = rand(0, 1);
